ISPN-857 - Favour Externalizer method implementations over @Marshalls
* @Marshalls annotation has been deprecated and the information it
provided is now provided by Externalizer implementations.
* Server core's CacheKey has been removed since it's no longer used
since ByteArrayKey was created.

+++ core/src/main/java/org/infinispan/marshall/Ids.java
@@ -22,7 +22,7 @@
 package org.infinispan.marshall;
 
 /**
- * Indexes for object types. These are currently limited to being unsigned bytes, so valid values are considered those
+ * Indexes for object types. These are currently limited to being unsigned ints, so valid values are considered those
  * in the range of 0 to 254. Please note that the use of 255 is forbidden since this is reserved for foreign, or user
  * defined, externalizers.
  *
@@ -33,50 +33,50 @@ public interface Ids {
    // No internal externalizer should use this upper limit Id or anything higher than that.
    static final int MAX_ID = 255;
 
-   static final byte ARRAY_LIST = 0;
-   static final byte LINKED_LIST = 1;
-   static final byte MAPS = 2;
-   static final byte JDK_SETS = 3;
-   static final byte SINGLETON_LIST = 4;
+   static final int ARRAY_LIST = 0;
+   static final int LINKED_LIST = 1;
+   static final int MAPS = 2;
+   static final int JDK_SETS = 3;
+   static final int SINGLETON_LIST = 4;
    // responses
-   static final byte SUCCESSFUL_RESPONSE = 5;
-   static final byte EXTENDED_RESPONSE = 6;
-   static final byte EXCEPTION_RESPONSE = 7;
-   static final byte UNSUCCESSFUL_RESPONSE = 8;
-   static final byte REQUEST_IGNORED_RESPONSE = 9;
+   static final int SUCCESSFUL_RESPONSE = 5;
+   static final int EXTENDED_RESPONSE = 6;
+   static final int EXCEPTION_RESPONSE = 7;
+   static final int UNSUCCESSFUL_RESPONSE = 8;
+   static final int REQUEST_IGNORED_RESPONSE = 9;
    // entries and values
-   static final byte IMMORTAL_ENTRY = 10;
-   static final byte MORTAL_ENTRY = 11;
-   static final byte TRANSIENT_ENTRY = 12;
-   static final byte TRANSIENT_MORTAL_ENTRY = 13;
-   static final byte IMMORTAL_VALUE = 14;
-   static final byte MORTAL_VALUE = 15;
-   static final byte TRANSIENT_VALUE = 16;
-   static final byte TRANSIENT_MORTAL_VALUE = 17;
+   static final int IMMORTAL_ENTRY = 10;
+   static final int MORTAL_ENTRY = 11;
+   static final int TRANSIENT_ENTRY = 12;
+   static final int TRANSIENT_MORTAL_ENTRY = 13;
+   static final int IMMORTAL_VALUE = 14;
+   static final int MORTAL_VALUE = 15;
+   static final int TRANSIENT_VALUE = 16;
+   static final int TRANSIENT_MORTAL_VALUE = 17;
    // internal collections (id=18 no longer in use, might get reused at a later stage)
-   static final byte IMMUTABLE_MAP = 19;
-   static final byte ATOMIC_HASH_MAP = 20;
+   static final int IMMUTABLE_MAP = 19;
+   static final int ATOMIC_HASH_MAP = 20;
    // others
-   static final byte GLOBAL_TRANSACTION = 38;
-   static final byte JGROUPS_ADDRESS = 39;
-   static final byte MARSHALLED_VALUE = 40;
-   static final byte TRANSACTION_LOG_ENTRY = 41;
-   static final byte BUCKET = 42;
-   static final byte DEADLOCK_DETECTING_GLOBAL_TRANSACTION = 43;
+   static final int GLOBAL_TRANSACTION = 38;
+   static final int JGROUPS_ADDRESS = 39;
+   static final int MARSHALLED_VALUE = 40;
+   static final int TRANSACTION_LOG_ENTRY = 41;
+   static final int BUCKET = 42;
+   static final int DEADLOCK_DETECTING_GLOBAL_TRANSACTION = 43;
    // 44 and 45 no longer in use, used to belong to tree module
-   static final byte ATOMIC_HASH_MAP_DELTA = 46;
-   static final byte ATOMIC_PUT_OPERATION = 47;
-   static final byte ATOMIC_REMOVE_OPERATION = 48;
-   static final byte ATOMIC_CLEAR_OPERATION = 49;
-   static final byte DEFAULT_CONSISTENT_HASH = 51;
-   static final byte UNION_CONSISTENT_HASH = 52;
-   static final byte UNSURE_RESPONSE = 54;
+   static final int ATOMIC_HASH_MAP_DELTA = 46;
+   static final int ATOMIC_PUT_OPERATION = 47;
+   static final int ATOMIC_REMOVE_OPERATION = 48;
+   static final int ATOMIC_CLEAR_OPERATION = 49;
+   static final int DEFAULT_CONSISTENT_HASH = 51;
+   static final int UNION_CONSISTENT_HASH = 52;
+   static final int UNSURE_RESPONSE = 54;
    // 55 - 56 no longer in use since server modules can now register their own externalizers
-   static final byte BYTE_ARRAY_KEY = 57;
+   static final int BYTE_ARRAY_KEY = 57;
    // 58 - 59 no longer in use since server modules can now register their own externalizers
-   static final byte NODE_TOPOLOGY_INFO = 60;
-   static final byte TOPOLOGY_AWARE_CH = 61;
+   static final int NODE_TOPOLOGY_INFO = 60;
+   static final int TOPOLOGY_AWARE_CH = 61;
    // commands (ids between 21 and 37 both inclusive and 50 and 53, are no longer in use, might get reused at a later stage)
-   static final byte REPLICABLE_COMMAND = 62;
+   static final int REPLICABLE_COMMAND = 62;
 
 }
